# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Usuários",
    "version": "1.0.0",
    "description": "API para registro, login, consulta e edição de usuários."
  },
  "paths": {
    "/api/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "400": { "description": "Erro de validação" }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Login do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Usuário ou senha inválidos" }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Listar usuários",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de usuários" },
          "401": { "description": "Token não informado ou inválido" }
        }
      }
    },
    "/api/users/{login}": {
      "put": {
        "summary": "Editar usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEdit"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Usuário editado" },
          "403": { "description": "Não autorizado" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserRegister": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "login": { "type": "string" },
          "senha": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": ["login", "senha", "email"]
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "login": { "type": "string" },
          "senha": { "type": "string" }
        },
        "required": ["login", "senha"]
      },
      "UserEdit": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "senha": { "type": "string" },
          "email": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}